requirements.txt
ultralytics
streamlit
pillow
opencv-python
numpy
roboflow